move to crdts utils, add tests'
