+++ packages/liveblocks-core/src/index.ts
@@ -29,7 +29,7 @@ export type {
   LiveObjectUpdate,
   StorageUpdate,
 } from "./crdts/StorageUpdates";
-export type { ToImmutable } from "./crdts/ToImmutable";
+export type { ToImmutable, toPlainLson } from "./crdts/utils";
 export {
   legacy_patchImmutableObject,
   lsonToJson,
@@ -49,12 +49,7 @@ export { asArrayWithLegacyMethods } from "./lib/LegacyArray";
 export { asPos, makePosition } from "./lib/position";
 export type { Resolve } from "./lib/Resolve";
 export { shallow } from "./lib/shallow";
-export {
-  b64decode,
-  isPlainObject,
-  toPlainLson,
-  tryParseJson,
-} from "./lib/utils";
+export { b64decode, isPlainObject, tryParseJson } from "./lib/utils";
 export type {
   AppOnlyAuthToken,
   AuthToken,
